rename ErrorType to InvalidFaceLicenseAlert
