+++ feature/setup/src/main/java/com/simprints/feature/setup/data/ErrorType.kt
@@ -40,7 +40,6 @@ internal enum class ErrorType(
     ),
     ;
 
-    @ExcludedFromGeneratedTestCoverageReports("Inner code of excluded file")
     fun toAlertArgs() = alertConfiguration {
         color = AlertColor.Gray
         title = this@ErrorType.customTitle
